remove required parsl configuration settings
